backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0

# GitHub
PyGithub==2.1.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# CORS
python-cors==1.0.0
